Strip symbols of Windows builds during install
Before this commit strip options were defined but not applied. For Mingw-w64
builds the non-stripped size is huge, about ten times larger than seems
necessary. This commit implements stripping of the symbol table during
install, for Windows builds only. Symbols can still be inspected in the dlls
wich are in the source tree.